move LambdaJavaTestlibsPackageInstaller to test module
