+++ tests/integration/awslambda/test_lambda_runtimes.py
@@ -1,10 +1,13 @@
 import json
 import os
 import shutil
+from typing import List
 
 import pytest
 
 from localstack.aws.api.lambda_ import Runtime
+from localstack.constants import LOCALSTACK_MAVEN_VERSION, MAVEN_REPO_URL
+from localstack.packages import DownloadInstaller, Package, PackageInstaller
 from localstack.services.awslambda.lambda_api import use_docker
 from localstack.services.awslambda.packages import lambda_java_libs_package
 from localstack.testing.aws.lambda_utils import is_old_provider
@@ -29,7 +32,33 @@ from tests.integration.awslambda.test_lambda import (
     THIS_FOLDER,
     read_streams,
 )
-from tests.packages.lambda_java_testlibs import lambda_java_testlibs_package
+
+# Java Test Jar Download (used for tests)
+TEST_LAMBDA_JAR_URL_TEMPLATE = "{url}/cloud/localstack/{name}/{version}/{name}-{version}-tests.jar"
+
+
+class LambdaJavaTestlibsPackage(Package):
+    def __init__(self):
+        super().__init__("JavaLambdaTestlibs", LOCALSTACK_MAVEN_VERSION)
+
+    def get_versions(self) -> List[str]:
+        return [LOCALSTACK_MAVEN_VERSION]
+
+    def _get_installer(self, version: str) -> PackageInstaller:
+        return LambdaJavaTestlibsPackageInstaller(version)
+
+
+class LambdaJavaTestlibsPackageInstaller(DownloadInstaller):
+    def __init__(self, version):
+        super().__init__("lambda-java-testlibs", version)
+
+    def _get_download_url(self) -> str:
+        return TEST_LAMBDA_JAR_URL_TEMPLATE.format(
+            version=self.version, url=MAVEN_REPO_URL, name="localstack-utils"
+        )
+
+
+lambda_java_testlibs_package = LambdaJavaTestlibsPackage()
 
 parametrize_python_runtimes = pytest.mark.parametrize("runtime", PYTHON_TEST_RUNTIMES)
 parametrize_node_runtimes = pytest.mark.parametrize("runtime", NODE_TEST_RUNTIMES)
